gibler color fix

+++ src/engine/utils/timer.rs
@@ -29,6 +29,6 @@ impl Timer {
     }
 
     pub fn get_as_percentage(&self) -> f32 {
-        return (self.get_current_time() as f32 - self.last_fired as f32) / self.timeout as f32;
+        return (self.get_current_time() - self.last_fired) as f32 / (self.timeout as f32);
     }
 }
